Move bookmark helpers to FolderManager. Deletes protocol extension.
